requirements.txt
langchain
langchain-core==0.3.22
langchain-openai==0.2.11
openai==1.57.0
python-dotenv==1.0.1
streamlit==1.31.1
tiktoken==0.8.0
google-generativeai==0.3.2
tqdm==4.67.1
pydantic==2.10.3
requests==2.32.3
plotly>=5.13.0
pandas>=1.5.3